fix(db): extract the entity inventory so config imports stay concrete
CI surfaced a genuine conflict my previous commit created between two specs:

  database.config.ts imports the ../entities BARREL
    -> agent's database.config.spec.ts PASSES (its barrel mock intercepts, so
       TypeORM never loads and the known path-scurry-under-Jest bug is avoided)
    -> api's trigger-internal.controller.spec.ts FAILS: the barrel forces the
       whole entity graph to evaluate mid-initialisation, ENTITIES ends up
       holding an undefined, and TypeOrmModule.forFeature(ENTITIES) dies with
       "A circular dependency has been detected inside @InjectRepository()"

  database.config.ts imports CONCRETE entity files
    -> api spec PASSES (cycle broken)
    -> agent spec FAILS: concrete imports bypass the barrel mock, TypeORM loads
       for real, and path-scurry throws
       "TypeError: Cannot read properties of undefined (reading 'native')"

Both requirements are legitimate, so satisfy both instead of picking a side.

The 98 entity imports + the ENTITIES array move to `database/_entities-inventory.ts`
(they were used for nothing else - verified: all 96 named imports appeared only
inside ENTITIES). database.config.ts imports ENTITIES from there and re-exports it,
so every existing importer is unaffected.

That keeps the imports concrete (cycle stays fixed) while giving the spec ONE
module to mock instead of a barrel - the same "one place to edit" spirit as
_entity-names.ts (EW-638). The spec's mock now synthesizes a distinct NAMED CLASS
per entity, because it asserts each ENTITIES entry is `typeof === 'function'` and
that all entries are unique.

Verified locally with the repo's pinned pnpm:
  apps/api    trigger-internal.{controller,module}.spec  2 suites, 16 tests PASS
  agent       database.{config,module}.spec + factory    3 suites, 81 tests PASS
  repo-wide   pnpm format:check                          clean
